requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
numpy==1.23.5
matplotlib==3.7.0
torch==2.0.0
reformer-pytorch==1.4.4
flask==3.1.0
requests==2.32.3
pyyaml==6.0.2
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.1
joblib==1.3.2
psutil==5.9.8
websockets==12.0
python-multipart==0.0.9
